pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "transform-service"
version = "0.1.0"
description = "Comprehensive and robust PySpark + Starburst transformation service"
authors = [{name="Your Name", email="you@example.com"}]
requires-python = ">=3.11,<3.14"
readme = "README.md"
license = { text = "Apache-2.0" }
keywords = ["pyspark", "starburst", "trino", "etl", "data-engineering"]
dependencies = [
  "pyspark==3.4.4",
  "PyYAML>=6.0.1",
  "pydantic>=2.0",
  "psutil>=5.9.0",
]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "License :: OSI Approved :: Apache Software License",
  "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://example.com"

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-cov>=5.0.0",
  "pytest-mock>=3.14.0",
  "flake8>=7.0.0",
  "mypy>=1.10.0",
  "types-PyYAML",
  "python-json-logger>=2.0",
]

[project.scripts]
transform-service = "transform_service.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["transform_service*"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --cov=transform_service --cov-report=term-missing"
testpaths = ["tests"]
filterwarnings = [
  "ignore::DeprecationWarning",
]

[tool.flake8]
max-line-length = 120
extend-ignore = ["E203", "W503"]

[tool.mypy]
python_version = "3.11"
warn_unused_ignores = true
warn_return_any = true
no_implicit_optional = true
ignore_missing_imports = true
